Demon Health and Hit
Added demon health from the health associated with the 'lights' - created an fsm template. Testing the death animation.
Cleaned up the naming conventions of some of the FSM that were not correct. This seems to be a stable build, accept the navmesh issue with the Deer Demon, which is slightly janky.